Rename and clean up test helper plugin
